Massive updates to the BASH command section!
As part of my work at Squarespace, I'm teaching from these notes again -- and it
turns out to be very helpful to have a reasonable list of BASH commands for
folks to learn. So! Here it is. Woo!

This adds more commands and examples of use for a whole heap of common Linux
commands!